requirements.txt
groq
pytesseract
opencv-python
pandas
PyYAML
watchdog
python-dotenv
transformers
torch
google-generativeai 